So apparently, the 'b.out' format that this code used was based on 32-bit longs, not 64-bit longs. Change the code back to 'int' which will work everywhere until we try to make it self-hosting on the venix box again (where we'll need long since ints are 16-bits there, longs are 32-bit). I'm deferring doing anything here because we may have a lot of type issues and I'd like to cope with them all later. I may revisit that....

+++ 8086/a86/a.out.h
@@ -4,14 +4,14 @@
  * Header prepended to each a.out file.
  */
 struct exec {
-	long	a_magic;	/* magic number */
-unsigned long	a_text;		/* size of text segment */
-unsigned long	a_data;		/* size of initialized data */
-unsigned long	a_bss;		/* size of uninitialized data */
-unsigned long	a_syms;		/* size of symbol table */
-unsigned long	a_entry;	/* entry point */
-unsigned long	a_trsize;	/* size of text relocation */
-unsigned long	a_drsize;	/* size of data relocation */
+	int	a_magic;	/* magic number */
+unsigned int	a_text;		/* size of text segment */
+unsigned int	a_data;		/* size of initialized data */
+unsigned int	a_bss;		/* size of uninitialized data */
+unsigned int	a_syms;		/* size of symbol table */
+unsigned int	a_entry;	/* entry point */
+unsigned int	a_trsize;	/* size of text relocation */
+unsigned int	a_drsize;	/* size of data relocation */
 };
 
 #define	OMAGIC	0407		/* old impure format */
